+++ src/main/java/com/familytree/service/FamilyService.java
@@ -1,5 +1,6 @@
 package com.familytree.service;
 
+
 public interface FamilyService {
     public void getDescendants(String personName);
 }
