--- src/main/java/com/familytree/service/FamilyServiceImpl.java
@@ -1,21 +1,0 @@
-package com.familytree.service;
-
-import com.familytree.data.DataStore;
-import com.familytree.data.Person;
-
-public class FamilyServiceImpl implements FamilyService {
-    private DataStore ds;
-
-    public FamilyServiceImpl(DataStore ds) {
-        this.ds = ds;
-    }
-
-    @Override
-    public void getDescendants(String personName) {
- 
-    }
-
-    private boolean hasChildren(Person person) {
-        return null != person.getChildren() &&  person.getChildren().length > 0;
-    }
-}
